Move refactored functions into new functions package and add some more tests
