+++ excellent/functions/functions.go
@@ -1,8 +1,7 @@
-package excellent
+package functions
 
 import (
 	"bytes"
-	"encoding/json"
 	"fmt"
 	"math"
 	"math/rand"
@@ -20,78 +19,80 @@ import (
 	"github.com/shopspring/decimal"
 )
 
-// Function defines the interface that Excellent functions must implement
-type Function func(env utils.Environment, args ...interface{}) interface{}
+var randSource = rand.NewSource(time.Now().UnixNano())
 
-// RegisterFunction registers a new function in Excellent
-func RegisterFunction(name string, function Function) {
-	FUNCTIONS[name] = function
+// XFunction defines the interface that Excellent functions must implement
+type XFunction func(env utils.Environment, args ...types.XValue) types.XValue
+
+// RegisterXFunction registers a new function in Excellent
+func RegisterXFunction(name string, function XFunction) {
+	XFUNCTIONS[name] = function
 }
 
-// FUNCTIONS is our map of functions available in Excellent which aren't tests
-var FUNCTIONS = map[string]Function{
-	"and": _And,
-	"if":  _If,
-	"or":  _Or,
+// XFUNCTIONS is our map of functions available in Excellent which aren't tests
+var XFUNCTIONS = map[string]XFunction{
+	"and": And,
+	"if":  If,
+	"or":  Or,
 
-	"length":  _Length,
-	"default": _Default,
-	"array":   _Array,
+	"length":  Length,
+	"default": Default,
+	"array":   Array,
 
-	"legacy_add": _LegacyAdd,
+	"legacy_add": LegacyAdd,
 
-	"round":      _Round,
-	"round_up":   _RoundUp,
-	"round_down": _RoundDown,
-	"max":        _Max,
-	"min":        _Min,
-	"mean":       _Mean,
-	"mod":        _Mod,
-	"rand":       _Rand,
-	"abs":        _Abs,
+	"round":      Round,
+	"round_up":   RoundUp,
+	"round_down": RoundDown,
+	"max":        Max,
+	"min":        Min,
+	"mean":       Mean,
+	"mod":        Mod,
+	"rand":       Rand,
+	"abs":        Abs,
 
-	"format_num": _FormatNum,
-	"read_code":  _ReadCode,
+	"format_num": FormatNum,
+	"read_code":  ReadCode,
 
-	"to_json":    _ToJSON,
-	"from_json":  _FromJSON,
-	"url_encode": _URLEncode,
+	"to_json":    ToJSON,
+	"from_json":  FromJSON,
+	"url_encode": URLEncode,
 
-	"char":              _Char,
-	"code":              _Code,
-	"split":             _Split,
-	"join":              _Join,
-	"title":             _Title,
-	"word":              _Word,
-	"remove_first_word": _RemoveFirstWord,
-	"word_count":        _WordCount,
-	"word_slice":        _WordSlice,
-	"field":             _Field,
-	"clean":             _Clean,
-	"left":              _Left,
-	"lower":             _Lower,
-	"right":             _Right,
-	"string_cmp":        _StringCmp,
-	"repeat":            _Repeat,
-	"replace":           _Replace,
-	"upper":             _Upper,
-	"percent":           _Percent,
+	"char":              Char,
+	"code":              Code,
+	"split":             Split,
+	"join":              Join,
+	"title":             Title,
+	"word":              Word,
+	"remove_first_word": RemoveFirstWord,
+	"word_count":        WordCount,
+	"word_slice":        WordSlice,
+	"field":             Field,
+	"clean":             Clean,
+	"left":              Left,
+	"lower":             Lower,
+	"right":             Right,
+	"string_cmp":        StringCmp,
+	"repeat":            Repeat,
+	"replace":           Replace,
+	"upper":             Upper,
+	"percent":           Percent,
 
-	"format_date":     _FormatDate,
-	"parse_date":      _ParseDate,
-	"date":            _Date,
-	"date_from_parts": _DateFromParts,
-	"date_diff":       _DateDiff,
-	"date_add":        _DateAdd,
-	"weekday":         _Weekday,
-	"tz":              _TZ,
-	"tz_offset":       _TZOffset,
-	"today":           _Today,
-	"now":             _Now,
-	"from_epoch":      _FromEpoch,
-	"to_epoch":        _ToEpoch,
+	"format_date":     FormatDate,
+	"parse_date":      ParseDate,
+	"date":            Date,
+	"date_from_parts": DateFromParts,
+	"date_diff":       DateDiff,
+	"date_add":        DateAdd,
+	"weekday":         Weekday,
+	"tz":              TZ,
+	"tz_offset":       TZOffset,
+	"today":           Today,
+	"now":             Now,
+	"from_epoch":      FromEpoch,
+	"to_epoch":        ToEpoch,
 
-	"format_urn": _FormatURN,
+	"format_urn": FormatURN,
 }
 
 //------------------------------------------------------------------------------------------
@@ -102,50 +103,50 @@ var FUNCTIONS = map[string]Function{
 // one of the parameters was a date or not. If one is a date, then the other side is
 // expected to be an integer with a number of days to add to the date, otherwise a normal
 // decimal addition is attempted.
-func _LegacyAdd(env utils.Environment, args ...interface{}) interface{} {
+func LegacyAdd(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 2 {
-		return fmt.Errorf("LEGACY_ADD requires exactly two arguments, got %d", len(args))
+		return types.NewXErrorf("LEGACY_ADD requires exactly two arguments, got %d", len(args))
 	}
 
 	// try to parse dates and decimals
-	date1, date1Err := types.ToDate(env, args[0])
-	date2, date2Err := types.ToDate(env, args[1])
+	date1, date1Err := types.ToXTime(env, args[0])
+	date2, date2Err := types.ToXTime(env, args[1])
 
-	dec1, dec1Err := types.ToDecimal(env, args[0])
-	dec2, dec2Err := types.ToDecimal(env, args[1])
+	dec1, dec1Err := types.ToXNumber(args[0])
+	dec2, dec2Err := types.ToXNumber(args[1])
 
 	// if they are both dates, that's an error
 	if date1Err == nil && date2Err == nil {
-		return fmt.Errorf("LEGACY_ADD cannot operate on two dates")
+		return types.NewXErrorf("LEGACY_ADD cannot operate on two dates")
 	}
 
 	// date and int, do a day addition
 	if date1Err == nil && dec2Err == nil {
-		if dec2.IntPart() < math.MinInt32 || dec2.IntPart() > math.MaxInt32 {
-			return fmt.Errorf("LEGACY_ADD cannot operate on integers greater than 32 bit")
+		if dec2.Native().IntPart() < math.MinInt32 || dec2.Native().IntPart() > math.MaxInt32 {
+			return types.NewXErrorf("LEGACY_ADD cannot operate on integers greater than 32 bit")
 		}
-		return date1.AddDate(0, 0, int(dec2.IntPart()))
+		return types.NewXTime(date1.Native().AddDate(0, 0, int(dec2.Native().IntPart())))
 	}
 
 	// int and date, do a day addition
 	if date2Err == nil && dec1Err == nil {
-		if dec1.IntPart() < math.MinInt32 || dec1.IntPart() > math.MaxInt32 {
-			return fmt.Errorf("LEGACY_ADD cannot operate on integers greater than 32 bit")
+		if dec1.Native().IntPart() < math.MinInt32 || dec1.Native().IntPart() > math.MaxInt32 {
+			return types.NewXErrorf("LEGACY_ADD cannot operate on integers greater than 32 bit")
 		}
-		return date2.AddDate(0, 0, int(dec1.IntPart()))
+		return types.NewXTime(date2.Native().AddDate(0, 0, int(dec1.Native().IntPart())))
 	}
 
 	// one of these doesn't look like a valid decimal either, bail
 	if dec1Err != nil {
-		return dec1Err
+		return types.NewXError(dec1Err)
 	}
 
 	if dec2Err != nil {
-		return dec2Err
+		return types.NewXError(dec2Err)
 	}
 
 	// normal decimal addition
-	return dec1.Add(dec2)
+	return types.NewXNumber(dec1.Native().Add(dec2.Native()))
 }
 
 //------------------------------------------------------------------------------------------
@@ -163,24 +164,18 @@ func _LegacyAdd(env utils.Environment, args ...interface{}) interface{} {
 //   @(length(1234)) -> ERROR
 //
 // @function length(object)
-func _Length(env utils.Environment, args ...interface{}) interface{} {
+func Length(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("LENGTH takes exactly one argument, got %d", len(args))
+		return types.NewXErrorf("LENGTH takes exactly one argument, got %d", len(args))
 	}
 
-	// argument must either be an object with length
+	// argument must be a value with length
 	lengthable, isLengthable := args[0].(types.Lengthable)
 	if isLengthable {
-		return decimal.New(int64(lengthable.Length()), 0)
+		return types.NewXNumberFromInt(lengthable.Length())
 	}
 
-	// or a string
-	asString, isString := args[0].(string)
-	if isString {
-		return decimal.New(int64(utf8.RuneCountInString(asString)), 0)
-	}
-
-	return fmt.Errorf("LENGTH requires an object with length as its first argument, got %s", reflect.TypeOf(args[0]))
+	return types.NewXErrorf("LENGTH requires an object with length as its first argument, got %s", reflect.TypeOf(args[0]))
 }
 
 // Default takes two arguments, returning `test` if not an error or nil, otherwise returning `default`
@@ -190,9 +185,9 @@ func _Length(env utils.Environment, args ...interface{}) interface{} {
 //   @(default(date("invalid-date"), "today")) -> today
 //
 // @function default(test, default)
-func _Default(env utils.Environment, args ...interface{}) interface{} {
+func Default(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 2 {
-		return fmt.Errorf("DEFAULT takes exactly two arguments, got %d", len(args))
+		return types.NewXErrorf("DEFAULT takes exactly two arguments, got %d", len(args))
 	}
 
 	// first argument is nil, return arg2
@@ -201,7 +196,7 @@ func _Default(env utils.Environment, args ...interface{}) interface{} {
 	}
 
 	// test whether arg1 is an error
-	_, isErr := args[0].(error)
+	_, isErr := args[0].(types.XError)
 	if isErr {
 		return args[1]
 	}
@@ -217,8 +212,8 @@ func _Default(env utils.Environment, args ...interface{}) interface{} {
 //   @(length(array("a", "b"))) -> "2"
 //
 // @function array(values...)
-func _Array(env utils.Environment, args ...interface{}) interface{} {
-	return types.NewArray(args...)
+func Array(env utils.Environment, args ...types.XValue) types.XValue {
+	return types.NewXArray(args...)
 }
 
 // FromJSON tries to parse `string` as JSON, returning a fragment you can index into
@@ -229,23 +224,14 @@ func _Array(env utils.Environment, args ...interface{}) interface{} {
 //   @(from_json("invalid json")) -> ERROR
 //
 // @function from_json(string)
-func _FromJSON(env utils.Environment, args ...interface{}) interface{} {
+func FromJSON(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("FROM_JSON takes exactly one string argument, got %d", len(args))
+		return types.NewXErrorf("FROM_JSON takes exactly one string argument, got %d", len(args))
 	}
 
-	arg, err := types.ToString(env, args[0])
-	if err != nil {
-		return err
-	}
+	arg := types.ToXString(args[0])
 
-	// unmarshal our string into a JSON fragment
-	var fragment types.JSONFragment
-	err = json.Unmarshal([]byte(arg), &fragment)
-	if err != nil {
-		return err
-	}
-	return fragment
+	return types.JSONToXValue([]byte(arg.Native()))
 }
 
 // ToJSON tries to return a JSON representation of `value`. An error is returned if there is
@@ -257,17 +243,12 @@ func _FromJSON(env utils.Environment, args ...interface{}) interface{} {
 //  @(to_json(now())) -> "2017-05-10T12:50:00.000000-07:00"
 //
 // @function to_json(value)
-func _ToJSON(env utils.Environment, args ...interface{}) interface{} {
+func ToJSON(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("TO_JSON takes exactly one argument, got %d", len(args))
+		return types.NewXErrorf("TO_JSON takes exactly one argument, got %d", len(args))
 	}
 
-	json, err := types.ToJSON(env, args[0])
-	if err != nil {
-		return err
-	}
-
-	return json
+	return types.ToXJSON(args[0])
 }
 
 // URLEncode URL encodes `string` for use in a URL parameter
@@ -276,17 +257,13 @@ func _ToJSON(env utils.Environment, args ...interface{}) interface{} {
 //  @(url_encode(10)) -> 10
 //
 // @function url_encode(string)
-func _URLEncode(env utils.Environment, args ...interface{}) interface{} {
+func URLEncode(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("URL_ENCODE takes exactly one argument, got %d", len(args))
+		return types.NewXErrorf("URL_ENCODE takes exactly one argument, got %d", len(args))
 	}
 
-	arg1, err := types.ToString(env, args[0])
-	if err != nil {
-		return err
-	}
-
-	return url.QueryEscape(arg1)
+	arg1 := types.ToXString(args[0])
+	return types.NewXString(url.QueryEscape(arg1.Native()))
 }
 
 //------------------------------------------------------------------------------------------
@@ -299,23 +276,17 @@ func _URLEncode(env utils.Environment, args ...interface{}) interface{} {
 //   @(and(true, false, true)) -> false
 //
 // @function and(tests...)
-func _And(env utils.Environment, args ...interface{}) interface{} {
+func And(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) == 0 {
-		return fmt.Errorf("AND requires at least one argument")
+		return types.NewXErrorf("AND requires at least one argument")
 	}
 
-	val, err := types.ToBool(env, args[0])
-	if err != nil {
-		return err
-	}
-	for _, iArg := range args[1:] {
-		iVal, err := types.ToBool(env, iArg)
-		if err != nil {
-			return err
+	for _, arg := range args {
+		if !types.ToXBool(arg) {
+			return types.XBoolFalse
 		}
-		val = val && iVal
 	}
-	return val
+	return types.XBoolTrue
 }
 
 // Or returns whether if any of the passed in arguments are truthy
@@ -324,24 +295,17 @@ func _And(env utils.Environment, args ...interface{}) interface{} {
 //   @(or(true, false, true)) -> true
 //
 // @function or(tests...)
-func _Or(env utils.Environment, args ...interface{}) interface{} {
+func Or(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) == 0 {
-		return fmt.Errorf("OR requires at least one argument")
+		return types.NewXErrorf("OR requires at least one argument")
 	}
 
-	val, err := types.ToBool(env, args[0])
-	if err != nil {
-		return err
-	}
-
-	for _, iArg := range args[1:] {
-		iVal, err := types.ToBool(env, iArg)
-		if err != nil {
-			return err
+	for _, arg := range args {
+		if types.ToXBool(arg) {
+			return types.XBoolTrue
 		}
-		val = val || iVal
 	}
-	return val
+	return types.XBoolFalse
 }
 
 // If evaluates the `test` argument, and if truthy returns `true_value`, if not returning `false_value`
@@ -352,20 +316,14 @@ func _Or(env utils.Environment, args ...interface{}) interface{} {
 //   @(if("foo" > "bar", "foo", "bar")) -> ERROR
 //
 // @function if(test, true_value, false_value)
-func _If(env utils.Environment, args ...interface{}) interface{} {
+func If(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 3 {
-		return fmt.Errorf("IF requires exactly 3 arguments, got %d", len(args))
+		return types.NewXErrorf("IF requires exactly 3 arguments, got %d", len(args))
 	}
 
-	truthy, err := types.ToBool(env, args[0])
-	if err != nil {
-		return err
-	}
-
-	if truthy {
+	if types.ToXBool(args[0]) {
 		return args[1]
 	}
-
 	return args[2]
 }
 
@@ -380,12 +338,12 @@ func _If(env utils.Environment, args ...interface{}) interface{} {
 //   @(abs("foo")) -> ERROR
 //
 // @function abs(num)
-func _Abs(env utils.Environment, args ...interface{}) interface{} {
-	dec, err := _checkOneDecimalArg(env, "ABS", args)
+func Abs(env utils.Environment, args ...types.XValue) types.XValue {
+	num, err := checkOneDecimalArg(env, "ABS", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
-	return dec.Abs()
+	return types.NewXNumber(num.Native().Abs())
 }
 
 // Round rounds `num` to the nearest value. You can optionally pass
@@ -401,25 +359,25 @@ func _Abs(env utils.Environment, args ...interface{}) interface{} {
 //   @(round("notnum", 2)) -> ERROR
 //
 // @function round(num [,places])
-func _Round(env utils.Environment, args ...interface{}) interface{} {
+func Round(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) < 1 || len(args) > 2 {
-		return fmt.Errorf("ROUND takes either one or two arguments")
+		return types.NewXErrorf("ROUND takes either one or two arguments")
 	}
 
-	dec, err := types.ToDecimal(env, args[0])
+	num, err := types.ToXNumber(args[0])
 	if err != nil {
-		return fmt.Errorf("ROUND's first argument must be decimal")
+		return types.NewXErrorf("ROUND's first argument must be decimal")
 	}
 
 	round := 0
 	if len(args) == 2 {
-		round, err = types.ToInt(env, args[1])
+		round, err = types.ToInteger(args[1])
 		if err != nil {
-			return fmt.Errorf("ROUND's decimal places argument must be integer")
+			return types.NewXErrorf("ROUND's decimal places argument must be integer")
 		}
 	}
 
-	return dec.Round(int32(round))
+	return types.NewXNumber(num.Native().Round(int32(round)))
 }
 
 // RoundUp rounds `num` up to the nearest integer value, also good at fighting weeds
@@ -429,13 +387,13 @@ func _Round(env utils.Environment, args ...interface{}) interface{} {
 //   @(round_up("foo")) -> ERROR
 //
 // @function round_up(num)
-func _RoundUp(env utils.Environment, args ...interface{}) interface{} {
-	dec, err := _checkOneDecimalArg(env, "ROUND_UP", args)
+func RoundUp(env utils.Environment, args ...types.XValue) types.XValue {
+	num, err := checkOneDecimalArg(env, "ROUND_UP", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return dec.Ceil()
+	return types.NewXNumber(num.Native().Ceil())
 }
 
 // RoundDown rounds `num` down to the nearest integer value
@@ -445,13 +403,13 @@ func _RoundUp(env utils.Environment, args ...interface{}) interface{} {
 //   @(round_down("foo")) -> ERROR
 //
 // @function round_down(num)
-func _RoundDown(env utils.Environment, args ...interface{}) interface{} {
-	dec, err := _checkOneDecimalArg(env, "ROUND_DOWN", args)
+func RoundDown(env utils.Environment, args ...types.XValue) types.XValue {
+	num, err := checkOneDecimalArg(env, "ROUND_DOWN", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return dec.Floor()
+	return types.NewXNumber(num.Native().Floor())
 }
 
 // Max takes a list of `values` and returns the greatest of them
@@ -461,23 +419,23 @@ func _RoundDown(env utils.Environment, args ...interface{}) interface{} {
 //   @(max(1, 10, "foo")) -> ERROR
 //
 // @function max(values...)
-func _Max(env utils.Environment, args ...interface{}) interface{} {
+func Max(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) == 0 {
-		return fmt.Errorf("MAX takes at least one argument")
+		return types.NewXErrorf("MAX takes at least one argument")
 	}
 
-	max, err := types.ToDecimal(env, args[0])
+	max, err := types.ToXNumber(args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	for _, v := range args[1:] {
-		val, err := types.ToDecimal(env, v)
+		val, err := types.ToXNumber(v)
 		if err != nil {
-			return err
+			return types.NewXError(err)
 		}
 
-		if val.Cmp(max) > 0 {
+		if val.Native().Cmp(max.Native()) > 0 {
 			max = val
 		}
 	}
@@ -491,23 +449,23 @@ func _Max(env utils.Environment, args ...interface{}) interface{} {
 //   @(min(1, 2, "foo")) -> ERROR
 //
 // @function min(values)
-func _Min(env utils.Environment, args ...interface{}) interface{} {
+func Min(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) == 0 {
-		return fmt.Errorf("MIN takes at least one argument")
+		return types.NewXErrorf("MIN takes at least one argument")
 	}
 
-	max, err := types.ToDecimal(env, args[0])
+	max, err := types.ToXNumber(args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	for _, v := range args[1:] {
-		val, err := types.ToDecimal(env, v)
+		val, err := types.ToXNumber(v)
 		if err != nil {
-			return err
+			return types.NewXError(err)
 		}
 
-		if val.Cmp(max) < 0 {
+		if val.Native().Cmp(max.Native()) < 0 {
 			max = val
 		}
 	}
@@ -521,22 +479,22 @@ func _Min(env utils.Environment, args ...interface{}) interface{} {
 //   @(mean(1, "foo")) -> ERROR
 //
 // @function mean(values)
-func _Mean(env utils.Environment, args ...interface{}) interface{} {
+func Mean(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) == 0 {
-		return fmt.Errorf("MEAN requires at least one argument, got 0")
+		return types.NewXErrorf("MEAN requires at least one argument, got 0")
 	}
 
 	sum := decimal.Zero
 
 	for _, val := range args {
-		dec, err := types.ToDecimal(env, val)
+		num, err := types.ToXNumber(val)
 		if err != nil {
-			return err
+			return types.NewXError(err)
 		}
-		sum = sum.Add(dec)
+		sum = sum.Add(num.Native())
 	}
 
-	return sum.Div(decimal.NewFromFloat(float64(len(args))))
+	return types.NewXNumber(sum.Div(decimal.New(int64(len(args)), 0)))
 }
 
 // Mod returns the remainder of the division of `divident` by `divisor`
@@ -546,16 +504,14 @@ func _Mean(env utils.Environment, args ...interface{}) interface{} {
 //   @(mod(5, "foo")) -> ERROR
 //
 // @function mod(dividend, divisor)
-func _Mod(env utils.Environment, args ...interface{}) interface{} {
-	arg1, arg2, err := _checkTwoDecimalArgs(env, "MOD", args)
+func Mod(env utils.Environment, args ...types.XValue) types.XValue {
+	num1, num2, err := checkTwoDecimalArgs(env, "MOD", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return arg1.Mod(arg2)
+	return types.NewXNumber(num1.Native().Mod(num2.Native()))
 }
-
-var _randSource = rand.NewSource(time.Now().UnixNano())
 
 // Rand returns either a single random decimal between 0-1 or a random integer between `floor` and `ceiling` (inclusive)
 //
@@ -563,38 +519,38 @@ var _randSource = rand.NewSource(time.Now().UnixNano())
 //  @(rand(1, 5)) == 3
 //
 // @function rand(floor, ceiling)
-func _Rand(env utils.Environment, args ...interface{}) interface{} {
+func Rand(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 0 && len(args) != 2 {
-		return fmt.Errorf("RAND takes either no arguments or two arguments, got %d", len(args))
+		return types.NewXErrorf("RAND takes either no arguments or two arguments, got %d", len(args))
 	}
 
 	if len(args) == 0 {
-		return decimal.NewFromFloat(rand.New(_randSource).Float64())
+		return types.NewXNumber(decimal.NewFromFloat(rand.New(randSource).Float64()))
 	}
 
-	min, err := types.ToDecimal(env, args[0])
+	min, err := types.ToXNumber(args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
-	max, err := types.ToDecimal(env, args[1])
+	max, err := types.ToXNumber(args[1])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	// turn to integers
-	min = min.Floor()
-	max = max.Floor()
+	minDec := min.Native().Floor()
+	maxDec := max.Native().Floor()
 
-	spread := min.Sub(max).Abs()
+	spread := minDec.Sub(maxDec).Abs()
 
 	// we add one here as the golang rand does is not inclusive, 2 will always return 1
 	// since our contract is inclusive of both ends we need one more
 	add := rand.New(randSource).Int63n(spread.IntPart() + 1)
 
-	if min.Cmp(max) <= 0 {
-		return min.Add(decimal.NewFromFloat(float64(add)))
+	if minDec.Cmp(maxDec) <= 0 {
+		return types.NewXNumber(minDec.Add(decimal.NewFromFloat(float64(add))))
 	}
-	return max.Add(decimal.NewFromFloat(float64(add)))
+	return types.NewXNumber(maxDec.Add(decimal.NewFromFloat(float64(add))))
 }
 
 // FormatNum returns `num` formatted with the passed in number of decimal `places` and optional `commas` dividing thousands separators
@@ -604,28 +560,25 @@ func _Rand(env utils.Environment, args ...interface{}) interface{} {
 //   @(format_num("foo", 2, false)) -> ERROR
 //
 // @function format_num(num, places, commas)
-func _FormatNum(env utils.Environment, args ...interface{}) interface{} {
+func FormatNum(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 3 {
-		return fmt.Errorf("FORMAT_NUM takes exactly three arguments, got %d", len(args))
+		return types.NewXErrorf("FORMAT_NUM takes exactly three arguments, got %d", len(args))
 	}
 
-	dec, err := types.ToDecimal(env, args[0])
+	num, err := types.ToXNumber(args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	places, err := types.ToInt(env, args[1])
+	places, err := types.ToInteger(args[1])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 	if places < 0 || places > 9 {
-		return fmt.Errorf("FORMAT_NUM must take 0-9 number of places, got %d", args[1])
+		return types.NewXErrorf("FORMAT_NUM must take 0-9 number of places, got %d", args[1])
 	}
 
-	commas, err := types.ToBool(env, args[2])
-	if err != nil {
-		return err
-	}
+	commas := types.ToXBool(args[2])
 
 	// build our format string
 	formatStr := bytes.Buffer{}
@@ -639,8 +592,8 @@ func _FormatNum(env utils.Environment, args ...interface{}) interface{} {
 			formatStr.WriteString("#")
 		}
 	}
-	f64, _ := dec.Float64()
-	return humanize.FormatFloat(formatStr.String(), f64)
+	f64, _ := num.Native().Float64()
+	return types.NewXString(humanize.FormatFloat(formatStr.String(), f64))
 }
 
 //------------------------------------------------------------------------------------------
@@ -657,21 +610,17 @@ func _FormatNum(env utils.Environment, args ...interface{}) interface{} {
 //   @(read_code("abcdef")) -> "a b c , d e f"
 //
 // @function read_code(code)
-func _ReadCode(env utils.Environment, args ...interface{}) interface{} {
+func ReadCode(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("READ_CODE takes exactly one argument, got %d", len(args))
+		return types.NewXErrorf("READ_CODE takes exactly one argument, got %d", len(args))
 	}
 
-	// convert to a string
-	val, err := types.ToString(env, args[0])
-	if err != nil {
-		return err
-	}
+	val := types.ToXString(args[0])
 
 	var output bytes.Buffer
 
 	// remove any leading +
-	val = strings.TrimLeft(val, "+")
+	val = types.NewXString(strings.TrimLeft(val.Native(), "+"))
 
 	length := len(val)
 
@@ -682,9 +631,9 @@ func _ReadCode(env utils.Environment, args ...interface{}) interface{} {
 			if i > 0 {
 				output.WriteString(" , ")
 			}
-			output.WriteString(strings.Join(strings.Split(val[i:i+3], ""), " "))
+			output.WriteString(strings.Join(strings.Split(val.Native()[i:i+3], ""), " "))
 		}
-		return output.String()
+		return types.NewXString(output.String())
 	}
 
 	// groups of four
@@ -693,9 +642,9 @@ func _ReadCode(env utils.Environment, args ...interface{}) interface{} {
 			if i > 0 {
 				output.WriteString(" , ")
 			}
-			output.WriteString(strings.Join(strings.Split(val[i:i+4], ""), " "))
+			output.WriteString(strings.Join(strings.Split(val.Native()[i:i+4], ""), " "))
 		}
-		return output.String()
+		return types.NewXString(output.String())
 	}
 
 	// default, just do one at a time
@@ -706,7 +655,7 @@ func _ReadCode(env utils.Environment, args ...interface{}) interface{} {
 		output.WriteRune(c)
 	}
 
-	return output.String()
+	return types.NewXString(output.String())
 }
 
 //------------------------------------------------------------------------------------------
@@ -723,18 +672,18 @@ func _ReadCode(env utils.Environment, args ...interface{}) interface{} {
 //   @(code(15)) -> "49"
 //
 // @function code(string)
-func _Code(env utils.Environment, args ...interface{}) interface{} {
-	str, err := _checkOneStringArg(env, "code", args)
+func Code(env utils.Environment, args ...types.XValue) types.XValue {
+	str, err := checkOneStringArg(env, "code", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	if len(str) == 0 {
-		return fmt.Errorf("CODE requires a string of at least one character")
+		return types.NewXErrorf("CODE requires a string of at least one character")
 	}
 
-	r, _ := utf8.DecodeRuneInString(str)
-	return int(r)
+	r, _ := utf8.DecodeRuneInString(str.Native())
+	return types.NewXNumberFromInt(int(r))
 }
 
 // Split splits `string` based on the passed in `delimeter`
@@ -748,27 +697,27 @@ func _Code(env utils.Environment, args ...interface{}) interface{} {
 //   @(split("a && b && c", " && ")) -> "a, b, c"
 //
 // @function split(string, delimeter)
-func _Split(env utils.Environment, args ...interface{}) interface{} {
+func Split(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 2 {
-		return fmt.Errorf("SPLIT takes exactly two arguments: string and delimiter, got %d", len(args))
+		return types.NewXErrorf("SPLIT takes exactly two arguments: string and delimiter, got %d", len(args))
 	}
 
 	s, err := types.ToString(env, args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	sep, err := types.ToString(env, args[1])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	splits := types.NewArray()
+	splits := types.NewXArray()
 
 	allSplits := strings.Split(s, sep)
 	for i := range allSplits {
 		if allSplits[i] != "" {
-			splits.Append(allSplits[i])
+			splits.Append(types.NewXString(allSplits[i]))
 		}
 	}
 	return splits
@@ -780,34 +729,28 @@ func _Split(env utils.Environment, args ...interface{}) interface{} {
 //   @(join(split("a.b.c", "."), " ")) -> "a b c"
 //
 // @function join(array, delimeter)
-func _Join(env utils.Environment, args ...interface{}) interface{} {
+func Join(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 2 {
-		return fmt.Errorf("JOIN takes exactly two arguments: the array to join and delimiter, got %d", len(args))
+		return types.NewXErrorf("JOIN takes exactly two arguments: the array to join and delimiter, got %d", len(args))
 	}
 
-	indexable, isIndexable := args[0].(types.Indexable)
+	indexable, isIndexable := args[0].(types.XIndexable)
 	if !isIndexable {
-		return fmt.Errorf("JOIN requires an indexable as its first argument, got %s", reflect.TypeOf(args[0]))
+		return types.NewXErrorf("JOIN requires an indexable as its first argument, got %s", reflect.TypeOf(args[0]))
 	}
 
-	sep, err := types.ToString(env, args[1])
-	if err != nil {
-		return err
-	}
+	sep := types.ToXString(args[1])
 
 	var output bytes.Buffer
 	for i := 0; i < indexable.Length(); i++ {
 		if i > 0 {
-			output.WriteString(sep)
+			output.WriteString(sep.Native())
 		}
-		itemAsStr, err := types.ToString(env, indexable.Index(i))
-		if err != nil {
-			return err
-		}
-		output.WriteString(itemAsStr)
+		itemAsStr := types.ToXString(indexable.Index(i))
+		output.WriteString(itemAsStr.Native())
 	}
 
-	return output.String()
+	return types.NewXString(output.String())
 }
 
 // Char returns the rune for the passed in codepoint, `num`, which may be unicode, this is the reverse of code
@@ -817,13 +760,18 @@ func _Join(env utils.Environment, args ...interface{}) interface{} {
 //   @(char("foo")) -> ERROR
 //
 // @function char(num)
-func _Char(env utils.Environment, args ...interface{}) interface{} {
-	arg, err := _checkOneDecimalArg(env, "CHAR", args)
+func Char(env utils.Environment, args ...types.XValue) types.XValue {
+	arg, err := checkOneDecimalArg(env, "CHAR", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return string(rune(arg.IntPart()))
+	num, err := types.ToInteger(arg)
+	if err != nil {
+		return types.NewXError(err)
+	}
+
+	return types.NewXString(string(rune(num)))
 }
 
 // Title titlecases the passed in `string`, capitalizing each word
@@ -833,13 +781,13 @@ func _Char(env utils.Environment, args ...interface{}) interface{} {
 //   @(title(123)) -> "123"
 //
 // @function title(string)
-func _Title(env utils.Environment, args ...interface{}) interface{} {
-	arg, err := _checkOneStringArg(env, "TITLE", args)
+func Title(env utils.Environment, args ...types.XValue) types.XValue {
+	arg, err := checkOneStringArg(env, "TITLE", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return strings.Title(arg)
+	return types.NewXString(strings.Title(arg.Native()))
 }
 
 // Word returns the word at the passed in `offset` for the passed in `string`
@@ -849,27 +797,24 @@ func _Title(env utils.Environment, args ...interface{}) interface{} {
 //   @(word("one two.three", 2)) -> "three"
 //
 // @function word(string, offset)
-func _Word(env utils.Environment, args ...interface{}) interface{} {
+func Word(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 2 {
-		return fmt.Errorf("WORD takes exactly two arguments, got %d", len(args))
+		return types.NewXErrorf("WORD takes exactly two arguments, got %d", len(args))
 	}
 
-	val, err := types.ToString(env, args[0])
+	val := types.ToXString(args[0])
+
+	word, err := types.ToInteger(args[1])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	word, err := types.ToInt(env, args[1])
-	if err != nil {
-		return err
-	}
-
-	words := utils.TokenizeString(val)
+	words := utils.TokenizeString(val.Native())
 	if word >= len(words) {
-		return fmt.Errorf("Word offset %d is greater than number of words %d", word, len(words))
+		return types.NewXErrorf("Word offset %d is greater than number of words %d", word, len(words))
 	}
 
-	return words[word]
+	return types.NewXString(words[word])
 }
 
 // RemoveFirstWord removes the 1st word of `string`
@@ -877,18 +822,18 @@ func _Word(env utils.Environment, args ...interface{}) interface{} {
 //   @(remove_first_word("foo bar")) -> "bar"
 //
 // @function remove_first_word(string)
-func _RemoveFirstWord(env utils.Environment, args ...interface{}) interface{} {
-	arg, err := _checkOneStringArg(env, "REMOVE_FIRST_WORD", args)
+func RemoveFirstWord(env utils.Environment, args ...types.XValue) types.XValue {
+	arg, err := checkOneStringArg(env, "REMOVE_FIRST_WORD", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	words := utils.TokenizeString(arg)
+	words := utils.TokenizeString(arg.Native())
 	if len(words) > 1 {
-		return strings.Join(words[1:], " ")
+		return types.NewXString(strings.Join(words[1:], " "))
 	}
 
-	return ""
+	return types.XStringEmpty
 }
 
 // WordSlice extracts a substring from `string` spanning from `start` up to but not-including `end`. (first word is 1)
@@ -898,30 +843,27 @@ func _RemoveFirstWord(env utils.Environment, args ...interface{}) interface{} {
 //   @(word_slice("foo bar", 3, 4)) -> ""
 //
 // @function word_slice(string, start, end)
-func _WordSlice(env utils.Environment, args ...interface{}) interface{} {
+func WordSlice(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 3 {
-		return fmt.Errorf("WORD_SLICE takes exactly three arguments, got %d", len(args))
+		return types.NewXErrorf("WORD_SLICE takes exactly three arguments, got %d", len(args))
 	}
 
-	arg, err := types.ToString(env, args[0])
-	if err != nil {
-		return fmt.Errorf("WORD_SLICE requires a string as its first argument")
-	}
+	arg := types.ToXString(args[0])
 
-	start, err := types.ToInt(env, args[1])
+	start, err := types.ToInteger(args[1])
 	if err != nil || start <= 0 {
-		return fmt.Errorf("WORD_SLICE must start with a positive index")
+		return types.NewXErrorf("WORD_SLICE must start with a positive index")
 	}
 	start--
 
-	stop, err := types.ToInt(env, args[2])
+	stop, err := types.ToInteger(args[2])
 	if err != nil || start < 0 {
-		return fmt.Errorf("WORD_SLICE must have a stop of 0 or greater")
+		return types.NewXErrorf("WORD_SLICE must have a stop of 0 or greater")
 	}
 
-	words := utils.TokenizeString(arg)
+	words := utils.TokenizeString(arg.Native())
 	if start >= len(words) {
-		return ""
+		return types.XStringEmpty
 	}
 
 	if stop >= len(words) {
@@ -929,9 +871,9 @@ func _WordSlice(env utils.Environment, args ...interface{}) interface{} {
 	}
 
 	if stop > 0 {
-		return strings.Join(words[start:stop], " ")
+		return types.NewXString(strings.Join(words[start:stop], " "))
 	}
-	return strings.Join(words[start:], " ")
+	return types.NewXString(strings.Join(words[start:], " "))
 }
 
 // WordCount returns the number of words in `string`
@@ -942,14 +884,14 @@ func _WordSlice(env utils.Environment, args ...interface{}) interface{} {
 //   @(word_count("😀😃😄😁")) -> 4
 //
 // @function word_count(string)
-func _WordCount(env utils.Environment, args ...interface{}) interface{} {
-	arg, err := _checkOneStringArg(env, "WORD_COUNT", args)
+func WordCount(env utils.Environment, args ...types.XValue) types.XValue {
+	arg, err := checkOneStringArg(env, "WORD_COUNT", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	words := utils.TokenizeString(arg)
-	return decimal.New(int64(len(words)), 0)
+	words := utils.TokenizeString(arg.Native())
+	return types.NewXNumberFromInt(len(words))
 }
 
 // Field splits `string` based on the passed in `delimiter` and returns the field at `offset`.  When splitting
@@ -963,33 +905,26 @@ func _WordCount(env utils.Environment, args ...interface{}) interface{} {
 //   @(field("a,b,c", "foo", ",")) -> ERROR
 //
 // @function field(string, offset, delimeter)
-func _Field(env utils.Environment, args ...interface{}) interface{} {
-	source, err := types.ToString(env, args[0])
+func Field(env utils.Environment, args ...types.XValue) types.XValue {
+	source := types.ToXString(args[0])
+	field, err := types.ToInteger(args[1])
 	if err != nil {
-		return err
-	}
-
-	field, err := types.ToInt(env, args[1])
-	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	if field < 0 {
-		return fmt.Errorf("Cannot use a negative index to FIELD")
+		return types.NewXErrorf("cannot use a negative index to FIELD")
 	}
 
-	sep, err := types.ToString(env, args[2])
-	if err != nil {
-		return err
-	}
+	sep := types.ToXString(args[2])
 
-	fields := strings.Split(source, sep)
+	fields := strings.Split(source.Native(), sep.Native())
 	if field >= len(fields) {
-		return ""
+		return types.XStringEmpty
 	}
 
 	// when using a space as a delimiter, we consider it splitting on whitespace, so remove empty values
-	if sep == " " {
+	if sep.Native() == " " {
 		var newFields []string
 		for _, field := range fields {
 			if field != "" {
@@ -999,7 +934,7 @@ func _Field(env utils.Environment, args ...interface{}) interface{} {
 		fields = newFields
 	}
 
-	return strings.TrimSpace(fields[field])
+	return types.NewXString(strings.TrimSpace(fields[field]))
 }
 
 // Clean strips any leading or trailing whitespace from `string``
@@ -1009,13 +944,13 @@ func _Field(env utils.Environment, args ...interface{}) interface{} {
 //   @(clean(123)) -> "123"
 //
 // @function clean(string)
-func _Clean(env utils.Environment, args ...interface{}) interface{} {
-	arg, err := _checkOneStringArg(env, "CLEAN", args)
+func Clean(env utils.Environment, args ...types.XValue) types.XValue {
+	arg, err := checkOneStringArg(env, "CLEAN", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return strings.TrimSpace(arg)
+	return types.NewXString(strings.TrimSpace(arg.Native()))
 }
 
 // Left returns the `len` most left characters of the passed in `string`
@@ -1026,16 +961,16 @@ func _Clean(env utils.Environment, args ...interface{}) interface{} {
 //   @(left("hello", -1)) -> ERROR
 //
 // @function left(string, len)
-func _Left(env utils.Environment, args ...interface{}) interface{} {
-	str, l, err := _checkOneStringOneIntArg(env, "LEFT", args)
+func Left(env utils.Environment, args ...types.XValue) types.XValue {
+	str, l, err := checkOneStringOneIntArg(env, "LEFT", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	// this weird construct does the right thing for multi-byte unicode
 	var output bytes.Buffer
 	i := 0
-	for _, r := range str {
+	for _, r := range str.Native() {
 		if i >= l {
 			break
 		}
@@ -1043,7 +978,7 @@ func _Left(env utils.Environment, args ...interface{}) interface{} {
 		i++
 	}
 
-	return output.String()
+	return types.NewXString(output.String())
 }
 
 // Lower lowercases the passed in `string`
@@ -1054,13 +989,13 @@ func _Left(env utils.Environment, args ...interface{}) interface{} {
 //   @(lower("😀")) -> "😀"
 //
 // @function lower(string)
-func _Lower(env utils.Environment, args ...interface{}) interface{} {
-	arg, err := _checkOneStringArg(env, "LOWER", args)
+func Lower(env utils.Environment, args ...types.XValue) types.XValue {
+	arg, err := checkOneStringArg(env, "LOWER", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return strings.ToLower(arg)
+	return types.NewXString(strings.ToLower(arg.Native()))
 }
 
 // Right returns the `len` most right characters of the passed in `string`
@@ -1071,25 +1006,25 @@ func _Lower(env utils.Environment, args ...interface{}) interface{} {
 //   @(right("hello", -1)) -> ERROR
 //
 // @function right(string, len)
-func _Right(env utils.Environment, args ...interface{}) interface{} {
-	str, l, err := _checkOneStringOneIntArg(env, "RIGHT", args)
+func Right(env utils.Environment, args ...types.XValue) types.XValue {
+	str, l, err := checkOneStringOneIntArg(env, "RIGHT", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	start := utf8.RuneCountInString(str) - l
+	start := utf8.RuneCountInString(str.Native()) - l
 
 	// this weird construct does the right thing for multi-byte unicode
 	var output bytes.Buffer
 	i := 0
-	for _, r := range str {
+	for _, r := range str.Native() {
 		if i >= start {
 			output.WriteRune(r)
 		}
 		i++
 	}
 
-	return output.String()
+	return types.NewXString(output.String())
 }
 
 // StringCmp returns the comparison between the strings `str1` and `str2`.
@@ -1101,13 +1036,13 @@ func _Right(env utils.Environment, args ...interface{}) interface{} {
 //   @(string_cmp("zzz", "aaa")) -> 1
 //
 // @function string_cmp(str1, str2)
-func _StringCmp(env utils.Environment, args ...interface{}) interface{} {
-	str1, str2, err := _checkTwoStringArgs(env, "STRING_CMP", args)
+func StringCmp(env utils.Environment, args ...types.XValue) types.XValue {
+	str1, str2, err := checkTwoStringArgs(env, "STRING_CMP", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return strings.Compare(str1, str2)
+	return types.NewXNumberFromInt(strings.Compare(str1.Native(), str2.Native()))
 }
 
 // Repeat return `string` repeated `count` number of times
@@ -1116,22 +1051,22 @@ func _StringCmp(env utils.Environment, args ...interface{}) interface{} {
 //   @(repeat("*", "foo")) -> ERROR
 //
 // @function repeat(string, count)
-func _Repeat(env utils.Environment, args ...interface{}) interface{} {
-	str, i, err := _checkOneStringOneIntArg(env, "REPEAT", args)
+func Repeat(env utils.Environment, args ...types.XValue) types.XValue {
+	str, i, err := checkOneStringOneIntArg(env, "REPEAT", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	if i < 0 {
-		return fmt.Errorf("REPEAT must be called with a positive integer, got %d", i)
+		return types.NewXErrorf("REPEAT must be called with a positive integer, got %d", i)
 	}
 
 	var output bytes.Buffer
 	for j := 0; j < i; j++ {
-		output.WriteString(str)
+		output.WriteString(str.Native())
 	}
 
-	return output.String()
+	return types.NewXString(output.String())
 }
 
 // Replace replaces all occurrences of `needle` with `replacement` in `string`
@@ -1140,27 +1075,16 @@ func _Repeat(env utils.Environment, args ...interface{}) interface{} {
 //   @(replace("foo bar", "baz", "zap")) -> "foo bar"
 //
 // @function replace(string, needle, replacement)
-func _Replace(env utils.Environment, args ...interface{}) interface{} {
+func Replace(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 3 {
-		return fmt.Errorf("REPLACE takes exactly three arguments, got %d", len(args))
+		return types.NewXErrorf("REPLACE takes exactly three arguments, got %d", len(args))
 	}
 
-	source, err := types.ToString(env, args[0])
-	if err != nil {
-		return err
-	}
+	source := types.ToXString(args[0])
+	find := types.ToXString(args[1])
+	replace := types.ToXString(args[2])
 
-	find, err := types.ToString(env, args[1])
-	if err != nil {
-		return err
-	}
-
-	replace, err := types.ToString(env, args[2])
-	if err != nil {
-		return err
-	}
-
-	return strings.Replace(source, find, replace, -1)
+	return types.NewXString(strings.Replace(source.Native(), find.Native(), replace.Native(), -1))
 }
 
 // Upper uppercases all characters in the passed `string`
@@ -1169,12 +1093,12 @@ func _Replace(env utils.Environment, args ...interface{}) interface{} {
 //   @(upper(123)) -> "123"
 //
 // @function upper(string)
-func _Upper(env utils.Environment, args ...interface{}) interface{} {
-	str, err := _checkOneStringArg(env, "UPPER", args)
+func Upper(env utils.Environment, args ...types.XValue) types.XValue {
+	str, err := checkOneStringArg(env, "UPPER", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
-	return strings.ToUpper(str)
+	return types.NewXString(strings.ToUpper(str.Native()))
 }
 
 // Percent converts `num` to a string represented as a percentage
@@ -1184,17 +1108,17 @@ func _Upper(env utils.Environment, args ...interface{}) interface{} {
 //   @(percent("foo")) -> ERROR
 //
 // @function percent(num)
-func _Percent(env utils.Environment, args ...interface{}) interface{} {
-	dec, err := _checkOneDecimalArg(env, "PERCENT", args)
+func Percent(env utils.Environment, args ...types.XValue) types.XValue {
+	num, err := checkOneDecimalArg(env, "PERCENT", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	// multiply by 100 and floor
-	percent := dec.Mul(decimal.NewFromFloat(100)).Round(0)
+	percent := num.Native().Mul(decimal.NewFromFloat(100)).Round(0)
 
 	// add on a %
-	return fmt.Sprintf("%d%%", percent.IntPart())
+	return types.NewXString(fmt.Sprintf("%d%%", percent.IntPart()))
 }
 
 //------------------------------------------------------------------------------------------
@@ -1243,47 +1167,38 @@ func _Percent(env utils.Environment, args ...interface{}) interface{} {
 //   @(parse_date("NOT DATE", "YYYY-MM-DD")) -> ERROR
 //
 // @function parse_date(string, format [,timezone])
-func _ParseDate(env utils.Environment, args ...interface{}) interface{} {
+func ParseDate(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) < 2 || len(args) > 3 {
-		return fmt.Errorf("PARSE_DATE requires at least two arguments, got %d", len(args))
-	}
-	arg1, err := types.ToString(env, args[0])
-	if err != nil {
-		return err
+		return types.NewXErrorf("PARSE_DATE requires at least two arguments, got %d", len(args))
 	}
 
-	format, err := types.ToString(env, args[1])
-	if err != nil {
-		return err
-	}
+	arg1 := types.ToXString(args[0])
+	format := types.ToXString(args[1])
 
 	// try to turn it to a go format
-	goFormat, err := utils.ToGoDateFormat(format)
+	goFormat, err := utils.ToGoDateFormat(format.Native())
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	// grab our location
 	location := env.Timezone()
 	if len(args) == 3 {
-		arg3, err := types.ToString(env, args[2])
-		if err != nil {
-			return err
-		}
+		arg3 := types.ToXString(args[2]).Native()
+
 		location, err = time.LoadLocation(arg3)
 		if err != nil {
-			return err
+			return types.NewXError(err)
 		}
 	}
 
 	// finally try to parse the date
-	parsed, err := time.ParseInLocation(goFormat, arg1, location)
+	parsed, err := time.ParseInLocation(goFormat, arg1.Native(), location)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	parsed = parsed.In(location)
-	return parsed
+	return types.NewXTime(parsed.In(location))
 }
 
 // FormatDate turns `date` into a string according to the `format` specified and in
@@ -1324,39 +1239,34 @@ func _ParseDate(env utils.Environment, args ...interface{}) interface{} {
 //   @(format_date("NOT DATE", "YYYY-MM-DD")) -> ERROR
 //
 // @function format_date(date, format [,timezone])
-func _FormatDate(env utils.Environment, args ...interface{}) interface{} {
+func FormatDate(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) < 1 || len(args) > 3 {
-		return fmt.Errorf("FORMAT_DATE takes one or two arguments, got %d", len(args))
+		return types.NewXErrorf("FORMAT_DATE takes one or two arguments, got %d", len(args))
 	}
 	date, err := types.ToDate(env, args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	format := fmt.Sprintf("%s %s", env.DateFormat().String(), env.TimeFormat().String())
 	if len(args) >= 2 {
-		format, err = types.ToString(env, args[1])
-		if err != nil {
-			return err
-		}
+		format = types.ToXString(args[1]).Native()
 	}
 
 	// try to turn it to a go format
 	goFormat, err := utils.ToGoDateFormat(format)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	// grab our location
 	location := env.Timezone()
 	if len(args) == 3 {
-		arg3, err := types.ToString(env, args[2])
-		if err != nil {
-			return err
-		}
+		arg3 := types.ToXString(args[2]).Native()
+
 		location, err = time.LoadLocation(arg3)
 		if err != nil {
-			return err
+			return types.NewXError(err)
 		}
 	}
 
@@ -1366,7 +1276,7 @@ func _FormatDate(env utils.Environment, args ...interface{}) interface{} {
 	}
 
 	// return the formatted date
-	return date.Format(goFormat)
+	return types.NewXString(date.Format(goFormat))
 }
 
 // Date turns `string` into a date according to the environment's settings
@@ -1378,21 +1288,19 @@ func _FormatDate(env utils.Environment, args ...interface{}) interface{} {
 //   @(date("NOT DATE")) -> ERROR
 //
 // @function date(string)
-func _Date(env utils.Environment, args ...interface{}) interface{} {
+func Date(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("DATE requires exactly one argument, got %d", len(args))
-	}
-	arg1, err := types.ToString(env, args[0])
-	if err != nil {
-		return err
+		return types.NewXErrorf("DATE requires exactly one argument, got %d", len(args))
 	}
 
-	date, err := utils.DateFromString(env, arg1)
+	arg1 := types.ToXString(args[0])
+
+	date, err := utils.DateFromString(env, arg1.Native())
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return date
+	return types.NewXTime(date)
 }
 
 // DateFromParts converts the passed in `year`, `month`` and `day`
@@ -1402,28 +1310,28 @@ func _Date(env utils.Environment, args ...interface{}) interface{} {
 //   @(date_from_parts(2017, 13, 15)) -> ERROR
 //
 // @function date_from_parts(year, month, day)
-func _DateFromParts(env utils.Environment, args ...interface{}) interface{} {
+func DateFromParts(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 3 {
-		return fmt.Errorf("DATE_FROM_PARTS requires three arguments, got %d", len(args))
+		return types.NewXErrorf("DATE_FROM_PARTS requires three arguments, got %d", len(args))
 	}
-	year, err := types.ToInt(env, args[0])
+	year, err := types.ToInteger(args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
-	month, err := types.ToInt(env, args[1])
+	month, err := types.ToInteger(args[1])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 	if month < 1 || month > 12 {
-		return fmt.Errorf("Invalidate value for month, must be 1-12")
+		return types.NewXErrorf("invalid value for month, must be 1-12")
 	}
 
-	day, err := types.ToInt(env, args[2])
+	day, err := types.ToInteger(args[2])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return time.Date(year, time.Month(month), day, 0, 0, 0, 0, env.Timezone())
+	return types.NewXTime(time.Date(year, time.Month(month), day, 0, 0, 0, 0, env.Timezone()))
 }
 
 // DateDiff returns the integer duration between `date1` and `date2` in the `unit` specified.
@@ -1436,55 +1344,45 @@ func _DateFromParts(env utils.Environment, args ...interface{}) interface{} {
 //   @(date_diff("2017-01-17", "2015-12-17", "y")) -> 2
 //
 // @function date_diff(date1, date2, unit)
-func _DateDiff(env utils.Environment, args ...interface{}) interface{} {
+func DateDiff(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 3 {
-		return fmt.Errorf("DATE_DIFF takes exactly three arguments, received %d", len(args))
+		return types.NewXErrorf("DATE_DIFF takes exactly three arguments, received %d", len(args))
 	}
 
-	date1, err := types.ToDate(env, args[0])
+	date1, err := types.ToXTime(env, args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	date2, err := types.ToDate(env, args[1])
+	date2, err := types.ToXTime(env, args[1])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	unit, err := types.ToString(env, args[2])
-	if err != nil {
-		return err
-	}
+	unit := types.ToXString(args[2])
 
 	// find the duration between our dates
-	duration := date1.Sub(date2)
+	duration := date1.Native().Sub(date2.Native())
 
 	// then convert based on our unit
-	switch unit {
-
+	switch unit.Native() {
 	case "s":
-		return int(duration / time.Second)
-
+		return types.NewXNumberFromInt(int(duration / time.Second))
 	case "m":
-		return int(duration / time.Minute)
-
+		return types.NewXNumberFromInt(int(duration / time.Minute))
 	case "h":
-		return int(duration / time.Hour)
-
+		return types.NewXNumberFromInt(int(duration / time.Hour))
 	case "d":
-		return utils.DaysBetween(date1, date2)
-
+		return types.NewXNumberFromInt(utils.DaysBetween(date1.Native(), date2.Native()))
 	case "w":
-		return int(utils.DaysBetween(date1, date2) / 7)
-
+		return types.NewXNumberFromInt(int(utils.DaysBetween(date1.Native(), date2.Native()) / 7))
 	case "M":
-		return utils.MonthsBetween(date1, date2)
-
+		return types.NewXNumberFromInt(utils.MonthsBetween(date1.Native(), date2.Native()))
 	case "y":
-		return date1.Year() - date2.Year()
+		return types.NewXNumberFromInt(date1.Native().Year() - date2.Native().Year())
 	}
 
-	return fmt.Errorf("Unknown unit: %s, must be one of s, m, h, D, W, M, Y", unit)
+	return types.NewXErrorf("Unknown unit: %s, must be one of s, m, h, D, W, M, Y", unit)
 }
 
 // DateAdd calculates the date value arrived at by adding `offset` number of `unit` to the `date`
@@ -1496,51 +1394,41 @@ func _DateDiff(env utils.Environment, args ...interface{}) interface{} {
 //   @(date_add("2017-01-15 10:45", 30, "m")) -> 2017-01-15T11:15:00.000000Z
 //
 // @function date_add(date, offset, unit)
-func _DateAdd(env utils.Environment, args ...interface{}) interface{} {
+func DateAdd(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 3 {
-		return fmt.Errorf("DATE_ADD takes exactly three arguments, received %d", len(args))
+		return types.NewXErrorf("DATE_ADD takes exactly three arguments, received %d", len(args))
 	}
 
-	date, err := types.ToDate(env, args[0])
+	date, err := types.ToXTime(env, args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	duration, err := types.ToInt(env, args[1])
+	duration, err := types.ToInteger(args[1])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	unit, err := types.ToString(env, args[2])
-	if err != nil {
-		return err
-	}
+	unit := types.ToXString(args[2])
 
-	switch unit {
-
+	switch unit.Native() {
 	case "s":
-		return date.Add(time.Duration(duration) * time.Second)
-
+		return types.NewXTime(date.Native().Add(time.Duration(duration) * time.Second))
 	case "m":
-		return date.Add(time.Duration(duration) * time.Minute)
-
+		return types.NewXTime(date.Native().Add(time.Duration(duration) * time.Minute))
 	case "h":
-		return date.Add(time.Duration(duration) * time.Hour)
-
+		return types.NewXTime(date.Native().Add(time.Duration(duration) * time.Hour))
 	case "d":
-		return date.AddDate(0, 0, duration)
-
+		return types.NewXTime(date.Native().AddDate(0, 0, duration))
 	case "w":
-		return date.AddDate(0, 0, duration*7)
-
+		return types.NewXTime(date.Native().AddDate(0, 0, duration*7))
 	case "M":
-		return date.AddDate(0, duration, 0)
-
+		return types.NewXTime(date.Native().AddDate(0, duration, 0))
 	case "y":
-		return date.AddDate(duration, 0, 0)
+		return types.NewXTime(date.Native().AddDate(duration, 0, 0))
 	}
 
-	return fmt.Errorf("Unknown unit: %s, must be one of s, m, h, d, w, M, y", unit)
+	return types.NewXErrorf("Unknown unit: %s, must be one of s, m, h, d, w, M, y", unit)
 }
 
 // Weekday returns the day of the week for `date`, 0 is sunday, 1 is monday..
@@ -1549,13 +1437,13 @@ func _DateAdd(env utils.Environment, args ...interface{}) interface{} {
 //   @(weekday("foo")) -> ERROR
 //
 // @function weekday(date)
-func _Weekday(env utils.Environment, args ...interface{}) interface{} {
-	date, err := _checkOneDateArg(env, "WEEKDAY", args)
+func Weekday(env utils.Environment, args ...types.XValue) types.XValue {
+	date, err := checkOneDateArg(env, "WEEKDAY", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return int(date.Weekday())
+	return types.NewXNumberFromInt(int(date.Native().Weekday()))
 }
 
 // TZ returns the timezone for `date``
@@ -1569,13 +1457,13 @@ func _Weekday(env utils.Environment, args ...interface{}) interface{} {
 //   @(tz("foo")) -> ERROR
 //
 // @function tz(date)
-func _TZ(env utils.Environment, args ...interface{}) interface{} {
-	date, err := _checkOneDateArg(env, "TZ", args)
+func TZ(env utils.Environment, args ...types.XValue) types.XValue {
+	date, err := checkOneDateArg(env, "TZ", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return date.Location().String()
+	return types.NewXString(date.Native().Location().String())
 }
 
 // TZOffset returns the offset for the timezone as a string +/- HHMM for `date`
@@ -1589,14 +1477,14 @@ func _TZ(env utils.Environment, args ...interface{}) interface{} {
 //   @(tz_offset("foo")) -> ERROR
 //
 // @function tz_offset(date)
-func _TZOffset(env utils.Environment, args ...interface{}) interface{} {
-	date, err := _checkOneDateArg(env, "TZ_OFFSET", args)
+func TZOffset(env utils.Environment, args ...types.XValue) types.XValue {
+	date, err := checkOneDateArg(env, "TZ_OFFSET", args)
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
 	// this looks like we are returning a set offset, but this is how go describes formats
-	return date.Format("-0700")
+	return types.NewXString(date.Native().Format("-0700"))
 
 }
 
@@ -1605,13 +1493,13 @@ func _TZOffset(env utils.Environment, args ...interface{}) interface{} {
 //  @(today()) -> 2017-01-20T00:00:00.000000Z
 //
 // @function today()
-func _Today(env utils.Environment, args ...interface{}) interface{} {
+func Today(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) > 0 {
-		return fmt.Errorf("TODAY takes no arguments, got %d", len(args))
+		return types.NewXErrorf("TODAY takes no arguments, got %d", len(args))
 	}
 
 	nowTZ := time.Now().In(env.Timezone())
-	return time.Date(nowTZ.Year(), nowTZ.Month(), nowTZ.Day(), 0, 0, 0, 0, env.Timezone())
+	return types.NewXTime(time.Date(nowTZ.Year(), nowTZ.Month(), nowTZ.Day(), 0, 0, 0, 0, env.Timezone()))
 }
 
 // FromEpoch returns a new date created from `num` which represents number of nanoseconds since January 1st, 1970 GMT
@@ -1619,17 +1507,17 @@ func _Today(env utils.Environment, args ...interface{}) interface{} {
 //   @(from_epoch(1497286619000000000)) -> 2017-06-12T16:56:59.000000Z
 //
 // @function from_epoch(num)
-func _FromEpoch(env utils.Environment, args ...interface{}) interface{} {
+func FromEpoch(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("FROM_EPOCH takes exactly one number argument, got %d", len(args))
+		return types.NewXErrorf("FROM_EPOCH takes exactly one number argument, got %d", len(args))
 	}
 
-	offset, err := types.ToDecimal(env, args[0])
+	offset, err := types.ToXNumber(args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return time.Unix(0, offset.IntPart()).In(env.Timezone())
+	return types.NewXTime(time.Unix(0, offset.Native().IntPart()).In(env.Timezone()))
 }
 
 // ToEpoch converts `date` to the number of nanoseconds since January 1st, 1970 GMT
@@ -1637,17 +1525,17 @@ func _FromEpoch(env utils.Environment, args ...interface{}) interface{} {
 //   @(to_epoch("2017-06-12T16:56:59.000000Z")) -> 1497286619000000000
 //
 // @function to_epoch(date)
-func _ToEpoch(env utils.Environment, args ...interface{}) interface{} {
+func ToEpoch(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("TO_EPOCH takes exactly one date argument, got %d", len(args))
+		return types.NewXErrorf("TO_EPOCH takes exactly one date argument, got %d", len(args))
 	}
 
-	date, err := types.ToDate(env, args[0])
+	date, err := types.ToXTime(env, args[0])
 	if err != nil {
-		return err
+		return types.NewXError(err)
 	}
 
-	return decimal.New(date.UnixNano(), 0)
+	return types.NewXNumberFromInt64(date.Native().UnixNano())
 }
 
 // Now returns the current date and time in the environment timezone
@@ -1655,12 +1543,12 @@ func _ToEpoch(env utils.Environment, args ...interface{}) interface{} {
 //  @(now()) -> 2017-01-20T15:35:65.153654Z
 //
 // @function now()
-func _Now(env utils.Environment, args ...interface{}) interface{} {
+func Now(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) > 0 {
-		return fmt.Errorf("NOW takes no arguments, got %d", len(args))
+		return types.NewXErrorf("NOW takes no arguments, got %d", len(args))
 	}
 
-	return time.Now().In(env.Timezone())
+	return types.NewXTime(time.Now().In(env.Timezone()))
 }
 
 //----------------------------------------------------------------------------------------
@@ -1679,129 +1567,106 @@ func _Now(env utils.Environment, args ...interface{}) interface{} {
 //   @(format_urn("NOT URN")) -> ERROR
 //
 // @function format_urn(urn)
-func _FormatURN(env utils.Environment, args ...interface{}) interface{} {
+func FormatURN(env utils.Environment, args ...types.XValue) types.XValue {
 	if len(args) != 1 {
-		return fmt.Errorf("FORMAT_URN takes one argument, got %d", len(args))
+		return types.NewXErrorf("FORMAT_URN takes one argument, got %d", len(args))
 	}
 
 	// if we've been passed an indexable like a URNList, use first item
 	urnArg := args[0]
 
-	indexable, isIndexable := urnArg.(types.Indexable)
+	indexable, isIndexable := urnArg.(types.XIndexable)
 	if isIndexable {
 		if indexable.Length() >= 1 {
 			urnArg = indexable.Index(0)
 		} else {
-			return ""
+			return types.XStringEmpty
 		}
 	}
 
-	urnString, err := types.ToString(env, urnArg)
-	if err != nil {
-		return err
-	}
+	urnString := types.ToXString(urnArg)
 
 	urn := urns.URN(urnString)
-	err = urn.Validate()
+	err := urn.Validate()
 	if err != nil {
-		return fmt.Errorf("%s is not a valid URN: %s", urnString, err)
+		return types.NewXErrorf("%s is not a valid URN: %s", urnString, err)
 	}
 
-	return urn.Format()
+	return types.NewXString(urn.Format())
 }
 
 //----------------------------------------------------------------------------------------
 // Utility Functions
 //----------------------------------------------------------------------------------------
 
-func _checkOneDecimalArg(env utils.Environment, funcName string, args []interface{}) (decimal.Decimal, error) {
+func checkOneDecimalArg(env utils.Environment, funcName string, args []types.XValue) (types.XNumber, error) {
 	if len(args) != 1 {
-		return decimal.Zero, fmt.Errorf("%s takes exactly one argument, got %d", funcName, len(args))
+		return types.XNumberZero, fmt.Errorf("%s takes exactly one argument, got %d", funcName, len(args))
 	}
 
-	arg1, err := types.ToDecimal(env, args[0])
+	arg1, err := types.ToXNumber(args[0])
 	if err != nil {
-		return decimal.Zero, err
+		return types.XNumberZero, err
 	}
 
 	return arg1, nil
 }
 
-func _checkOneStringArg(env utils.Environment, funcName string, args []interface{}) (string, error) {
+func checkOneStringArg(env utils.Environment, funcName string, args []types.XValue) (types.XString, error) {
 	if len(args) != 1 {
-		return "", fmt.Errorf("%s takes exactly one argument, got %d", funcName, len(args))
+		return types.XStringEmpty, fmt.Errorf("%s takes exactly one argument, got %d", funcName, len(args))
 	}
 
-	arg1, err := types.ToString(env, args[0])
-	if err != nil {
-		return "", err
-	}
-
-	return arg1, nil
+	return types.ToXString(args[0]), nil
 }
 
-func _checkTwoStringArgs(env utils.Environment, funcName string, args []interface{}) (string, string, error) {
+func checkTwoStringArgs(env utils.Environment, funcName string, args []types.XValue) (types.XString, types.XString, error) {
 	if len(args) != 2 {
-		return "", "", fmt.Errorf("%s takes exactly two string arguments, got %d", funcName, len(args))
+		return types.XStringEmpty, types.XStringEmpty, fmt.Errorf("%s takes exactly two string arguments, got %d", funcName, len(args))
 	}
 
-	arg1, err := types.ToString(env, args[0])
-	if err != nil {
-		return "", "", err
-	}
-
-	arg2, err := types.ToString(env, args[1])
-	if err != nil {
-		return "", "", err
-	}
-
-	return arg1, arg2, nil
+	return types.ToXString(args[0]), types.ToXString(args[1]), nil
 }
 
-func _checkOneStringOneIntArg(env utils.Environment, funcName string, args []interface{}) (string, int, error) {
+func checkOneStringOneIntArg(env utils.Environment, funcName string, args []types.XValue) (types.XString, int, error) {
 	if len(args) != 2 {
-		return "", 0, fmt.Errorf("%s takes exactly two arguments, got %d", funcName, len(args))
+		return types.XStringEmpty, 0, fmt.Errorf("%s takes exactly two arguments, got %d", funcName, len(args))
 	}
 
-	arg1, err := types.ToString(env, args[0])
+	num, err := types.ToInteger(args[1])
 	if err != nil {
-		return "", 0, err
+		return types.XStringEmpty, 0, err
 	}
 
-	arg2, err := types.ToInt(env, args[1])
-	if err != nil {
-		return "", 0, err
-	}
-
-	return arg1, arg2, err
+	return types.ToXString(args[0]), num, err
 }
 
-func _checkTwoDecimalArgs(env utils.Environment, funcName string, args []interface{}) (decimal.Decimal, decimal.Decimal, error) {
+func checkTwoDecimalArgs(env utils.Environment, funcName string, args []types.XValue) (types.XNumber, types.XNumber, error) {
 	if len(args) != 2 {
-		return decimal.Zero, decimal.Zero, fmt.Errorf("%s takes exactly two arguments, got %d", funcName, len(args))
+		return types.XNumberZero, types.XNumberZero, fmt.Errorf("%s takes exactly two arguments, got %d", funcName, len(args))
 	}
 
-	arg1, err := types.ToDecimal(env, args[0])
+	num1, err := types.ToXNumber(args[0])
 	if err != nil {
-		return decimal.Zero, decimal.Zero, err
+		return types.XNumberZero, types.XNumberZero, err
 	}
 
-	arg2, err := types.ToDecimal(env, args[1])
+	num2, err := types.ToXNumber(args[1])
 	if err != nil {
-		return decimal.Zero, decimal.Zero, err
+		return types.XNumberZero, types.XNumberZero, err
 	}
 
-	return arg1, arg2, nil
+	return num1, num2, nil
 }
 
-func _checkOneDateArg(env utils.Environment, funcName string, args []interface{}) (time.Time, error) {
+func checkOneDateArg(env utils.Environment, funcName string, args []types.XValue) (types.XTime, error) {
 	if len(args) != 1 {
-		return utils.ZeroTime, fmt.Errorf("%s takes exactly one argument, got %d", funcName, len(args))
+		return types.XTimeZero, fmt.Errorf("%s takes exactly one argument, got %d", funcName, len(args))
 	}
 
-	arg1, err := types.ToDate(env, args[0])
+	arg1, err := types.ToXTime(env, args[0])
 	if err != nil {
-		return utils.ZeroTime, err
+		return types.XTimeZero, err
 	}
 
 	return arg1, err
